requirements.txt
neo4j==5.28.1,
pandas==2.1.4,
numpy==1.26.4,
matplotlib==3.10.1,
seaborn==0.13.2,
python-dotenv==1.1.0,
typing-extensions>=4.0.0,
pytest==8.3.5,
pytest-cov==6.1.1,
black==25.1.0,
flake8==7.2.0,
mypy==1.15.0,
isort==6.0.1,
pylint==3.3.6,
dash==2.9.3,
dash-bootstrap-components==1.7.1,
dash-cytoscape==1.0.2,
plotly==5.24.1,
pyvis==0.3.2,
bokeh==3.7.2,
holoviews==1.20.2,
plotly-calplot==0.1.20,

